Move spec helpers out of System
Turns out these apply equally well to request specs
