+++ spec/support/omniauth_helper.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "omniauth_test_response_builder"
+
+module OmniAuthHelper
+  OmniAuth.config.test_mode = true
+  OmniAuth.config.logger = Logger.new("/dev/null")
+
+  def stub_auth_hash(**)
+    OmniauthTestResponseBuilder.new.stub_auth_hash(**)
+  end
+end
